Modified methods working with API and added "human" models
